backend/requirements.txt
django
pillow
whitenoise
social-auth-app-django
djangorestframework
